feat: Update dependencies and add Vercel analytics
Updates project dependencies to their latest versions, including @google/genai, lucide-react, @supabase/supabase-js, and react-markdown.

Introduces Vercel analytics to track application usage and user engagement.

Refactors the Habit type definition to include counter-based tracking, progress tracking, and target dates, enhancing habit tracking capabilities. Also updates the emoji list for a wider selection.

+++ components/JournalSection.tsx
@@ -80,7 +80,7 @@ const JournalSection: React.FC<JournalSectionProps> = ({ journals, setJournals, 
 
   const formatTimestamp = (isoStr: string) => {
     const date = new Date(isoStr);
-    return date.toLocaleDateString('en-US', { month: 'short', day: 'numeric', year: 'numeric' });
+    return date.toLocaleDateString('en-US', { month: 'long', day: 'numeric', year: 'numeric' });
   };
 
   // Grouping Logic
